Fixed multiple issues with Sunpaths / Celestial Sphere...
- Made the time field on the second background page non-selectable.
- Made items behind the background window inaccessible.
- Fixed day of year input problems (see steps below).
- Fixed southern hemisphere negative sign issue (see steps below).
- Removed the explanation window to avoid a serious bug (which is also on website version; see notes below).

Steps to fix the day of year input problems:
- Delete redundant Flash UI components from Library: "Globe Simulator / Flash UI Components".
- Apply Patch 2019-1 for the combobox.
- Add "_global._popUpLevel = 31000;" in FComboBoxClass constructor so it opens above the background page.
- Fix typo in the "Day of Year Selector" symbol. Search for "dayOfYear>=365" and replace with "dayOfYearZB>=365".
- Open FUIComponent and find "executeCallBack". Change to "executeCallback".

Steps to fix the southern hemisphere negative sign issue:
- Open "Latitude Selector and Text Inputs".
- Find "this.latitudeSlider.value = Math.abs(this._latitude);" in the update function and remove the "Math.abs".
- Replace the if/else blocks in onLatitudeSliderChanged with just "this._latitude = this.latitudeSlider.value;".

Notes on explanation window bug:
Manipulating a celestial sphere, grading the problem, and then opening the explanation window will cause the movie to enter an infinite loop, freezing the player. This also happens in the online version. (It gets stuck in the first while loop in param_Set in include_dynamic.txt. Specifically, the line "if (itemParamArray.itemValue == paramArray[i2].itemValue) {" continuously evaluates to true.) It seems this problem manifests only when a sphere has been manipulated first. Given the seriousness of the problem, the fact that it hasn't been reported before, and the challenge it would be to fix, I decided just to disable the explanation window in the first frame setup ("explanationMode = false;").